Move Address code to address.js
Exports OrderOverview so state of Order can be accessed elsewhere
