+++ assets/js/admin/orders/order-overview/index.js
@@ -1,5 +1,3 @@
-/* global _, $ */
-
 /**
  * Internal dependencies
  */
@@ -13,13 +11,9 @@ import { State } from './models/state.js';
 // Temporarily include old Refund flow.
 import './_refund.js';
 
-/**
- * Setup Order Overview on DOM ready.
- *
- * @since 3.0
- */
-jQueryReady( () => {
-	// Do nothing if no data is available for hydration.
+let overview;
+
+( () => {
 	if ( ! window.eddAdminOrderOverview ) {
 		return;
 	}
@@ -54,7 +48,7 @@ jQueryReady( () => {
 	} );
 
 	// Create Overview.
-	const overview = new Overview( {
+	overview = new Overview( {
 		state,
 	} );
 
@@ -82,83 +76,6 @@ jQueryReady( () => {
 			...adjustment,
 		} )
 	} );
+} ) ();
 
-	// ... finally render the Overview once all data is set.
-	overview.render();
-
-	/**
-	 * Adjusts Overview tax configuration when the Customer's address changes.
-	 *
-	 * @since 3.0
-	 */
-	( () => {
-		const countryInput = document.getElementById(
-			'edd_order_address_country'
-		);
-		const regionInput = document.getElementById(
-			'edd_order_address_region'
-		);
-
-		if ( ! ( countryInput && regionInput ) ) {
-			return;
-		}
-
-		/**
-		 * Retrieves a tax rate based on the currently selected Address.
-		 *
-		 * @since 3.0
-		 */
-		function getTaxRate() {
-			const country = $( '#edd_order_address_country' ).val();
-			const region = $( '#edd_order_address_region' ).val();
-
-			const nonce = document.getElementById( 'edd_get_tax_rate_nonce' )
-				.value;
-
-			wp.ajax.send( 'edd_get_tax_rate', {
-				data: {
-					nonce,
-					country,
-					region,
-				},
-				/**
-				 * Updates the Overview's tax configuration on successful retrieval.
-				 *
-				 * @since 3.0
-				 *
-				 * @param {Object} response AJAX response.
-				 */
-				success( response ) {
-					let { tax_rate: rate } = response;
-
-					// Make a percentage.
-					rate = rate * 100;
-
-					overview.options.state.set( 'hasTax', {
-						country,
-						region,
-						rate,
-					} );
-				},
-				/*
-				 * Updates the Overview's tax configuration on failed retrieval.
-				 *
-				 * @since 3.0
-				 */
-				error() {
-					overview.options.state.set( 'hasTax', false );
-				},
-			} );
-		}
-
-		// Update rate on Address change.
-		//
-		// Wait for Region field to be replaced when Country changes.
-		// Wait for typing when Regino field changes.
-		// jQuery listeners for Chosen compatibility.
-		$( '#edd_order_address_country' ).on( 'change', _.debounce( getTaxRate, 250 ) );
-
-		$( '#edd-order-address' ).on( 'change', '#edd_order_address_region', getTaxRate );
-		$( '#edd-order-address' ).on( 'keyup', '#edd_order_address_region', _.debounce( getTaxRate, 250 ) );
-	} )();
-} );
+export default overview;
